added security to Api access

+++ app/Http/Controllers/NewsController.php
@@ -8,39 +8,49 @@ use App\News;
 
 class NewsController extends Controller
 {
-    public function all()
+    public function all(Request $request)
     {
-        return News::all();
+        if($request->header("key") == "ESISAAPIACCESS"){
+            return News::all();
+        }
+        return response(401, 401);
     }
 
     public function add(Request $request)
     {
-        $news = new News;
-        if (isset($request->title) && isset($request->content)) {
-            $news->title = $request->title;
-            $news->content = $request->content;
-            $news->postDate = date('Y-m-d');
-            if(isset($request->file)){
-                $f = $request->file('file');
-                $resFile = basename($f->getClientOriginalName(), '.' . $f->getClientOriginalExtension()) . '_' .date('Y_m_d_H_i_s') . '.' . $f->getClientOriginalExtension();
-                $newPath = $f->storeAs("public/newsFiles", $resFile);
-                $newPath;
-                $news->filePath = $resFile;
+        if($request->header("key") == "ESISANEWSADMIN"){
+            
+            $news = new News;
+            if (isset($request->title) && isset($request->content)) {
+                $news->title = $request->title;
+                $news->content = $request->content;
+                $news->postDate = date('Y-m-d');
+                if(isset($request->file)){
+                    $f = $request->file('file');
+                    $resFile = basename($f->getClientOriginalName(), '.' . $f->getClientOriginalExtension()) . '_' .date('Y_m_d_H_i_s') . '.' . $f->getClientOriginalExtension();
+                    $newPath = $f->storeAs("public/newsFiles", $resFile);
+                    $newPath;
+                    $news->filePath = $resFile;
+                }
+                $news->save();
+                //$news = News::create($request->all());
+                return response()->json($news, 201);            
+            }else{
+                return 'error';
             }
-            $news->save();
-            //$news = News::create($request->all());
-            return response()->json($news, 201);            
-        }else{
-            return 'error';
         }
+        return response(401, 401);
     }
 
     public function remove(Request $request, $id)
     {
-        $news = News::findOrFail($id);
-        Storage::delete("public/newsFiles/" . $news->filePath);
-        $news->delete();
+        if($request->header("key") == "ESISANEWSADMIN"){
+            $news = News::findOrFail($id);
+            Storage::delete("public/newsFiles/" . $news->filePath);
+            $news->delete();
 
-        return 204;
+            return 204;
+        }
+        return response(401, 401);
     }
 }
